передаю привет создателям сайта se.ifmo.ru которые каждые 5 минут меняют условия лабы

+++ src/main/java/olechka/lab5/models/Coordinates.java
@@ -13,11 +13,11 @@ public class Coordinates {
     @HumanDescription(format = "дробное число")
     private double y; //Поле не может быть null
 
-    public double getX() {
+    public Integer getX() {
         return x;
     }
 
-    public Double getY() {
+    public double getY() {
         return y;
     }
